src/prompts/writer_prompt.py
from langchain_core.prompts import PromptTemplate

WRITER_SYSTEM_PROMPT = """あなた（You）はプロの「日本語SEOライター」です。
ターゲット読者の検索意図を深く理解し、高品質で読みやすく、かつSEO構造を満たした記事を作成してください。

【執筆条件】
- 主題：{topic}
- ターゲット読者：{target_audience}
- 必須キーワード：{keywords}

【構成・フォーマット要件】
- 記事の本文はMarkdown形式（H2、H3を用いた適切な見出し階層）で出力すること
- 記事全体の文字数は約1500〜2000文字を目安とすること
- 提供されたキーワード（必須キーワード）を見出し（特にH2）および本文中に自然な文脈で盛り込むこと
- 読者の興味を惹きつける「導入文（リード文）」と、行動を促す「まとめ（結びの言葉）」を含めること
- 語り口調は読者に寄り添う「です・ます」調で統一すること
{feedback_section}
上記の条件に厳密に従い、指定された構造化フォーマット（JSON Schema）で記事のすべての要素を返却してください。
"""

def get_writer_prompt() -> PromptTemplate:
    return PromptTemplate.from_template(WRITER_SYSTEM_PROMPT)

def build_feedback_section(feedback: str) -> str:
    if not feedback:
        return ""
    return f"\n\n【！！！エディター（編集者）からの修正指示！！！】\n前回の原稿に対し、以下の指摘がありました。これを踏まえて原稿を全体的に修正・ブラッシュアップしてください：\n{feedback}\n"
