pyproject.toml
[project]
name = "seo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.23",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
